move code to more logical place
