--- src/vs/base/browser/ui/aria/ariaDelegate.ts
@@ -1,8 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { IDisposable } from 'vs/base/common/lifecycle';
-
-export interface IScopedAccessibilityProgressSignalDelegate extends IDisposable { }
